requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
openai
PyPDF2>=3.0.0
python-dotenv>=1.0.1